Agrego nuevos endpoints de ROADMAP

+++ servicio-deposito/src/main/java/com/backend/tpi_backend/servicio_deposito/model/Ubicacion.java
@@ -14,7 +14,7 @@ public class Ubicacion {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id; // 🔹 Cambiado de Long a Integer
 
     @Column(nullable = false, length = 150)
     private String direccion;
